pyproject.toml
[tool.poetry]
name = "activity-tracker-drf-project"
version = "0.1.0"
description = ""
authors = ["avdeevaa <tess-avdeeva@seznam.cz>"]
readme = "README.md"
packages = [{include = "activity_tracker_drf_project"}]

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.2"
psycopg2-binary = "^2.9.9"
pillow = "^10.2.0"
djangorestframework = "^3.14.0"
django-filter = "^23.5"
djangorestframework-simplejwt = "^5.3.1"
coverage = "^7.4.1"
drf-yasg = "^1.21.7"
stripe = "^8.0.0"
celery = "^5.3.6"
redis = "^5.0.1"
eventlet = "^0.35.1"
django-celery-beat = "^2.1.0"
django-cors-headers = "^4.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
